+++ integration_tests/contracts/contracts/TestICA.sol
@@ -15,7 +15,7 @@ contract TestICA {
         SUCCESS,
         FAIL
     }
-    mapping (uint64 => Status) public statusMap;
+    mapping (string => mapping (uint64 => Status)) public statusMap;
     event OnPacketResult(uint64 seq, Status status);
 
     function encodeRegister(string memory connectionID, string memory version) internal view returns (bytes memory) {
@@ -102,14 +102,18 @@ contract TestICA {
         return lastSeq;
     }
 
-    function onPacketResultCallback(uint64 seq, bool ack) external payable returns (bool) {
+    function getStatus(string calldata packetSrcChannel, uint64 seq) public view returns (Status) {
+        return statusMap[packetSrcChannel][seq];
+    }
+
+    function onPacketResultCallback(uint64 seq, bool ack, string calldata packetSrcChannel, string calldata packetDstChannel) external payable returns (bool) {
         // To prevent called by arbitrary user
         require(msg.sender == module_address);
         Status status = Status.FAIL;
         if (ack) {
             status = Status.SUCCESS;
         }
-        statusMap[seq] = status;
+        statusMap[packetSrcChannel][seq] = status;
         emit OnPacketResult(seq, status);
         return true;
     }